pyproject.toml
[tool.black]
line-length = 119

[tool.poetry]
name = "zero-shot-captcha-solver"
version = "0.1.3"
description = ""
authors = ["Ludwig Stumpp <42147848+LudwigStumpp@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "zero_shot_captcha_solver"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.5"
Pillow = "^9.3.0"
streamlit-image-select = "^0.5.1"
transformers = "^4.28.1"
torch = "^2.0.0"
streamlit = "^1.22.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.17.1"
ipywidgets = "^8.0.2"
pre-commit = "^2.20.0"
mypy = "^0.991"
black = "^22.10.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
